feat: use different approach to make iframeapi
