+++ src/components/spotify-iframe.tsx
@@ -1,47 +1,15 @@
-import { increaseSongPlayCount } from '@/apis/songs';
 import React from 'react';
+import type { SpotifyController, IFrameAPI } from '@/hooks/useiFrameAPI';
+import { increaseSongPlayCount } from '@/apis/songs';
 
-type SpotifyController = {
-  addListener: (event: string, callback: (data: unknown) => void) => void;
-  removeListener: (event: string) => void;
-};
-
-export const SpotifyIframe = ({ id }: { id: string }) => {
+export const SpotifyIframe = ({ id, iFrameAPI }: {
+  id: string
+  iFrameAPI?: IFrameAPI
+}) => {
   const embedRef = React.useRef(null);
   const spotifyEmbedControllerRef = React.useRef<SpotifyController | null>(null);
-  const [iFrameAPI, setIFrameAPI] = React.useState<{
-    createController: (
-    element: HTMLElement | null,
-    options: { width: string; height: string; uri: string },
-    callback: (controller: SpotifyController) => void
-    ) => void;
-  } | undefined>(undefined);
   const [playerLoaded, setPlayerLoaded] = React.useState(false);
   const uri = `spotify:track:${id}`;
-
-  React.useEffect(() => {
-    if (document.querySelector('script[src="https://open.spotify.com/embed/iframe-api/v1"]')) {
-      return;
-    }
-    const script = document.createElement('script');
-    script.src = 'https://open.spotify.com/embed/iframe-api/v1';
-    script.async = true;
-    document.body.appendChild(script);
-    return () => {
-      document.body.removeChild(script);
-    };
-  }, []);
-
-  React.useEffect(() => {
-    if (iFrameAPI) {
-      return;
-    }
-
-    // @ts-expect-error Spotify uses global callback
-    window.onSpotifyIframeApiReady = (SpotifyIframeApi) => {
-      setIFrameAPI(SpotifyIframeApi);
-    };
-  }, [iFrameAPI]);
 
   React.useEffect(() => {
     if (playerLoaded || iFrameAPI === undefined) {
