+++ src/routes/feed/index.tsx
@@ -1,6 +1,7 @@
 import { useGetFeed } from '@/hooks/useFeed';
 import Card from './components/Card';
 import { Tabs, TabsList, TabsTrigger } from '@/components/tabs';
+import { useIFrameAPI } from '@/hooks/useiFrameAPI';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 
 export default function Feed () {
@@ -8,6 +9,7 @@ export default function Feed () {
   const [searchParams] = useSearchParams();
   const currentSort = searchParams.get('sortBy') || 'createdAt';
   const { data: feeds } = useGetFeed(currentSort as 'createdAt' | 'likeCount');
+  const iFrameAPI = useIFrameAPI();
 
   const handleTabChange = (value: string) => {
     if (value === 'createdAt') {
@@ -29,7 +31,7 @@ export default function Feed () {
       </div>
       <div className='flex flex-col gap-2'>
         {feeds.map((item) => (
-          <Card key={item.id} item={item} />
+          <Card key={item.id} item={item} iFrameAPI={iFrameAPI} />
         ))}
       </div>
     </div>
