+++ src/routes/profile/index.tsx
@@ -6,6 +6,7 @@ import Card from '@/routes/feed/components/Card';
 import more from '@/assets/more.svg';
 import { LogOut } from 'lucide-react';
 import { useNavigate } from 'react-router-dom';
+import { useIFrameAPI } from '@/hooks/useiFrameAPI';
 
 export default function Profile () {
   const { data: user } = useGetUser();
@@ -14,6 +15,7 @@ export default function Profile () {
   const { data: calendar } = useCalendar();
   const { data: userLikes } = useGetUserLikes();
   const navigate = useNavigate();
+  const iFrameAPI = useIFrameAPI();
 
   const handleLogout = () => {
     localStorage.clear();
@@ -55,14 +57,14 @@ export default function Profile () {
           ? (
             <>
               {calendar?.map((item) => (
-                <Card key={item.id} item={item} isProfile />
+                <Card key={item.id} item={item} isProfile iFrameAPI={iFrameAPI} />
               ))}
             </>
             )
           : (
             <>
               {userLikes?.map((item) => (
-                <Card key={item.id} item={item} />
+                <Card key={item.id} item={item} iFrameAPI={iFrameAPI} />
               ))}
             </>
             )}
